Converting No Sites View to View Controller. Also, launching Add Site action sheet correctly when 'Add new site' button pressed.
