--- WordPress/Classes/ViewRelated/NUX/NoSitesView.xib
@@ -1,83 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="13771" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="landscape">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="13772"/>
-        <capability name="Aspect ratio constraints" minToolsVersion="5.1"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" layoutMarginsFollowReadableWidth="YES" id="iN0-l3-epB" customClass="NoSitesView" customModule="WordPress" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="460" height="377"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <imageView userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="mysites-nosites" translatesAutoresizingMaskIntoConstraints="NO" id="G4e-dB-lxv" userLabel="No Sites Image">
-                    <rect key="frame" x="55" y="-60" width="350" height="219"/>
-                    <constraints>
-                        <constraint firstAttribute="width" secondItem="G4e-dB-lxv" secondAttribute="height" multiplier="8:5" id="4VW-jo-3yu"/>
-                        <constraint firstAttribute="width" constant="350" id="K6h-zU-akj"/>
-                    </constraints>
-                </imageView>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Create a new site for your business, magazine, or personal blog; or connect an existing WordPress installation." textAlignment="center" lineBreakMode="wordWrap" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontForContentSizeCategory="YES" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="ofN-jE-Mju">
-                    <rect key="frame" x="10" y="109" width="440" height="204"/>
-                    <fontDescription key="fontDescription" style="UICTFontTextStyleHeadline"/>
-                    <color key="textColor" red="0.18039215689999999" green="0.2274509804" blue="0.32549019610000002" alpha="1" colorSpace="calibratedRGB"/>
-                    <nil key="highlightedColor"/>
-                </label>
-                <button opaque="NO" contentMode="scaleToFill" verticalCompressionResistancePriority="749" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="TN9-gF-7ik" customClass="LoginButton" customModule="WordPress" customModuleProvider="target">
-                    <rect key="frame" x="180.5" y="333" width="99" height="44"/>
-                    <constraints>
-                        <constraint firstAttribute="height" constant="44" id="1Ip-Te-zAW"/>
-                    </constraints>
-                    <fontDescription key="fontDescription" style="UICTFontTextStyleBody"/>
-                    <state key="normal" title="Add new site">
-                        <color key="titleColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                    </state>
-                    <userDefinedRuntimeAttributes>
-                        <userDefinedRuntimeAttribute type="boolean" keyPath="isPrimary" value="YES"/>
-                    </userDefinedRuntimeAttributes>
-                    <connections>
-                        <action selector="addSiteButtonPressed:" destination="iN0-l3-epB" eventType="touchUpInside" id="WDU-57-U5x"/>
-                    </connections>
-                </button>
-            </subviews>
-            <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-            <constraints>
-                <constraint firstItem="ofN-jE-Mju" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="20" id="3JW-Wy-mfD">
-                    <variation key="heightClass=compact" constant="10"/>
-                </constraint>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="bottom" secondItem="TN9-gF-7ik" secondAttribute="bottom" id="3o0-sK-2a5"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="ofN-jE-Mju" secondAttribute="trailing" constant="20" id="J8o-sK-kBH">
-                    <variation key="heightClass=compact" constant="10"/>
-                </constraint>
-                <constraint firstItem="G4e-dB-lxv" firstAttribute="centerX" secondItem="vUN-kp-3ea" secondAttribute="centerX" id="Zpp-hn-0cs"/>
-                <constraint firstItem="G4e-dB-lxv" firstAttribute="top" secondItem="vUN-kp-3ea" secondAttribute="top" constant="-60" id="b6V-mf-GGH"/>
-                <constraint firstItem="TN9-gF-7ik" firstAttribute="centerX" secondItem="vUN-kp-3ea" secondAttribute="centerX" id="bxr-C3-MkT"/>
-                <constraint firstItem="ofN-jE-Mju" firstAttribute="top" secondItem="G4e-dB-lxv" secondAttribute="bottom" constant="-30" id="dbC-wC-Oob">
-                    <variation key="heightClass=compact" constant="-50"/>
-                </constraint>
-                <constraint firstItem="TN9-gF-7ik" firstAttribute="top" secondItem="ofN-jE-Mju" secondAttribute="bottom" constant="30" id="epL-QN-XvF">
-                    <variation key="heightClass=compact" constant="20"/>
-                </constraint>
-            </constraints>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <viewLayoutGuide key="safeArea" id="vUN-kp-3ea"/>
-            <connections>
-                <outlet property="addSiteButton" destination="TN9-gF-7ik" id="gbO-0C-3si"/>
-                <outlet property="noSitesTitle" destination="ofN-jE-Mju" id="TYP-bF-hKC"/>
-                <outlet property="titleLeadingConstraint" destination="3JW-Wy-mfD" id="EkL-c6-HlJ"/>
-                <outlet property="titleTrailingConstraint" destination="J8o-sK-kBH" id="heS-I7-Wr6"/>
-            </connections>
-            <point key="canvasLocation" x="-462" y="-130.5"/>
-        </view>
-    </objects>
-    <resources>
-        <image name="mysites-nosites" width="400" height="250"/>
-    </resources>
-</document>
